+++ flask_typed_routes/fields.py
@@ -6,11 +6,13 @@ input fields of the API.
 import abc
 import collections
 import inspect
+import typing as t
 
 import flask
 import pydantic.fields
 
 import flask_typed_routes.errors as ftr_errors
+import flask_typed_routes.utils as ftr_utils
 
 Unset = object()
 Undef = pydantic.fields.PydanticUndefined
@@ -31,18 +33,15 @@ class Field(abc.ABC):
     """
 
     kind = None
+    multi_types = (list, t.List, tuple, t.Tuple, set, t.Set, frozenset, t.FrozenSet) # noqa UP006
 
-    __slots__ = ("embed", "multi", "field_info", "annotation", "name")
+    __slots__ = ("embed", "field_info", "annotation", "name")
 
-    def __init__(self, *args, embed=False, multi=False, **kwargs):
+    def __init__(self, *args, embed=False, **kwargs):
         if embed and self.kind != FieldTypes.body:
             raise ftr_errors.InvalidParameterTypeError("Only 'Body' fields can be embedded.")
-        if multi and self.kind not in (FieldTypes.header, FieldTypes.cookie, FieldTypes.query):
-            msg = "Only 'Header', 'Cookie' and 'Query' fields can have multiple values."
-            raise ftr_errors.InvalidParameterTypeError(msg)
 
         self.embed = embed  # `Body` fields can be embedded
-        self.multi = multi  # `Header`, `Cookie` and `Query` fields can have multiple values
         self.field_info = pydantic.fields.Field(*args, **kwargs)
         # These attributes are set by the `flask_typed_routes.core.parse_field` function
         self.annotation = None
@@ -78,8 +77,29 @@ class Field(abc.ABC):
     def is_required(self):
         return self.default == inspect.Parameter.empty
 
-    def fetch(self, data):
+    @classmethod
+    def is_multi_field(cls, annotation):
+        """Check if the annotation is a multi-value type."""
+
+        if annotation:
+            tp = t.get_args(annotation)[0] if ftr_utils.is_annotated(annotation) else annotation
+            return tp in cls.multi_types or t.get_origin(tp) in cls.multi_types
+        else:
+            return False
+
+    def fetch_single_value(self, data):
         return data.get(self.alias, Unset) if self.alias else data
+
+    def fetch_model_value(self, obj):
+        data = dict()
+        for name, info in self.annotation.model_fields.items():
+            alias = info.alias or name
+            if alias in obj:
+                if self.is_multi_field(info.annotation):
+                    data[alias] = obj.getlist(alias)
+                else:
+                    data[alias] = obj.get(alias)
+        return data
 
 
 class Path(Field):
@@ -87,7 +107,7 @@ class Path(Field):
 
     @property
     def value(self):
-        return self.fetch(flask.request.view_args)
+        return self.fetch_single_value(flask.request.view_args)
 
 
 class Query(Field):
@@ -95,7 +115,11 @@ class Query(Field):
 
     @property
     def value(self):
-        return self.fetch(flask.request.args.to_dict(flat=not self.multi))
+        if ftr_utils.is_subclass(self.annotation, pydantic.BaseModel):
+            return self.fetch_model_value(flask.request.args)
+        else:
+            flat = not self.is_multi_field(self.annotation)
+            return self.fetch_single_value(flask.request.args.to_dict(flat=flat))
 
 
 class Cookie(Field):
@@ -103,7 +127,11 @@ class Cookie(Field):
 
     @property
     def value(self):
-        return self.fetch(flask.request.cookies.to_dict(flat=not self.multi))
+        if ftr_utils.is_subclass(self.annotation, pydantic.BaseModel):
+            return self.fetch_model_value(flask.request.cookies)
+        else:
+            flat = not self.is_multi_field(self.annotation)
+            return self.fetch_single_value(flask.request.cookies.to_dict(flat=flat))
 
 
 class Header(Field):
@@ -111,14 +139,18 @@ class Header(Field):
 
     @property
     def value(self):
-        items = flask.request.headers.items()
-        if self.multi:
-            data = collections.defaultdict(list)
-            for k, v in items:
-                data[k].append(v)
+        if ftr_utils.is_subclass(self.annotation, pydantic.BaseModel):
+            return self.fetch_model_value(flask.request.headers)
         else:
-            data = dict(items)
-        return self.fetch(data)
+            items = flask.request.headers.items()
+            multi = self.is_multi_field(self.annotation)
+            if multi:
+                data = collections.defaultdict(list)
+                for k, v in items:
+                    data[k].append(v)
+            else:
+                data = dict(items)
+            return self.fetch_single_value(data)
 
 
 class Body(Field):
@@ -126,4 +158,4 @@ class Body(Field):
 
     @property
     def value(self):
-        return self.fetch(flask.request.json or dict())
+        return self.fetch_single_value(flask.request.json or dict())