+++ flask_typed_routes/utils.py
@@ -89,7 +89,7 @@ def pretty_errors(fields, errors, /):
     :rtype: list[dict]
     """
 
-    locators = {(field.locator): field for field in fields}
+    locators = {field.locator: field for field in fields}
     for error in errors:
         loc = list(error["loc"])
         field = locators[loc[0]]
@@ -107,13 +107,13 @@ def extract_rule_params(rule, /):
     return frozenset(rule_regex.findall(rule))
 
 
-def get_func_path(func):
+def get_func_path(func, /):
     """Get the full path of a function/method."""
 
     return f"{func.__module__}.{func.__qualname__}"
 
 
-def cleandoc(func):
+def cleandoc(func, /):
     docstring = func.__doc__
     return inspect.cleandoc(docstring) if docstring else ""