+++ tests/unit/test_fields.py
@@ -1,6 +1,8 @@
+import typing
 import unittest.mock as mock
 
 import flask
+import pydantic
 import pytest
 import werkzeug.test
 import werkzeug.wrappers
@@ -9,40 +11,85 @@ import flask_typed_routes.errors as ftr_errors
 import flask_typed_routes.fields as ftr_fields
 
 
+class MyListFieldModel(pydantic.BaseModel):
+    list_field: typing.Annotated[list[str], pydantic.Field(alias='list-field')]
+
+
 def test_path_field(flask_app_auto):
     with flask_app_auto.test_request_context('/'):
         flask.request.view_args = {'id': 123}
         assert ftr_fields.Path(alias='id').value == 123
 
 
-@pytest.mark.parametrize('multi', [True, False])
-def test_query_field(flask_app_auto, multi):
-    expected = ['term'] if multi else 'term'
-    with flask_app_auto.test_request_context('/?search=term'):
-        result = ftr_fields.Query(alias='search', multi=multi).value
-        assert result == expected
+@pytest.mark.parametrize(
+    'url, annotation, expected',
+    [
+        ('/?search=term1&search=term2', list[str], ['term1', 'term2']),
+        ('/?search=term1&search=term2', str, 'term1'),
+        ('/?search=term1,term2', list[str], ['term1,term2']),
+        ('/?search=term1,term2', str, 'term1,term2'),
+    ]
+)
+def test_query_standalone_field(flask_app_auto, url, annotation, expected):
+    with flask_app_auto.test_request_context(url):
+        field = ftr_fields.Query(alias='search')
+        field.annotation = annotation
+        assert field.value == expected
 
 
-@pytest.mark.parametrize('multi', [True, False])
-def test_cookie_field(flask_app_auto, multi):
-    expected = ['abc123', 'abc124'] if multi else 'abc123'
+@pytest.mark.parametrize('url, expected', [
+    ('/?list-field=term1&list-field=term2', {'list-field': ['term1', 'term2']}),
+    ('/?list-field=term1,term2', {'list-field': ['term1,term2']}),
+])
+def test_query_model_list_field(flask_app_auto, url, expected):
+    with flask_app_auto.test_request_context(url):
+        field = ftr_fields.Query()
+        field.annotation = MyListFieldModel
+        assert field.value == expected
+
+
+@pytest.mark.parametrize(
+    'cookie, annotation, expected',
+    [
+        ('session_id=abc123; session_id=abc124', list[str], ['abc123', 'abc124']),
+        ('session_id=abc123; session_id=abc124', str, 'abc123'),
+        ('session_id=abc123', list[str], ['abc123']),
+        ('session_id=abc123', str, 'abc123'),
+    ]
+)
+def test_cookie_standalone_field(flask_app_auto, cookie, annotation, expected):
     with flask_app_auto.test_request_context('/'):
-        headers = [('Cookie', 'session_id=abc123; session_id=abc124')]
+        headers = [('Cookie', cookie)]
         builder = werkzeug.test.EnvironBuilder(path='/', headers=headers)
         env = builder.get_environ()
         request = werkzeug.wrappers.Request(env)
         with mock.patch('flask.request', request):
-            result = ftr_fields.Cookie(alias='session_id', multi=multi).value
+            field = ftr_fields.Cookie(alias='session_id')
+            field.annotation = annotation
+            assert field.value == expected
 
-        assert result == expected
+
+@pytest.mark.parametrize('cookie, expected', [
+    ('list-field=abc123; list-field=abc124', {'list-field': ['abc123', 'abc124']}),
+    ('list-field=abc123,abc124', {'list-field': ['abc123,abc124']}),
+])
+def test_cookie_model_list_field(flask_app_auto, cookie, expected):
+    with flask_app_auto.test_request_context('/'):
+        headers = [('Cookie', cookie)]
+        builder = werkzeug.test.EnvironBuilder(path='/', headers=headers)
+        env = builder.get_environ()
+        request = werkzeug.wrappers.Request(env)
+        with mock.patch('flask.request', request):
+            field = ftr_fields.Cookie()
+            field.annotation = MyListFieldModel
+            assert field.value == expected
 
 
-def test_header_field(flask_app_auto):
+def test_header_standalone_field(flask_app_auto):
     expected = 'Bearer token1'
     headers = [
         ('Authorization', 'Bearer token1'),
     ]
-    # Fixed: The `headers` multi dont work
     with flask_app_auto.test_request_context('/', headers=headers):
         result = ftr_fields.Header(alias='Authorization').value
         assert result == expected
@@ -66,7 +113,23 @@ def test_bad_embed_field(field_class):
         field_class(embed=True)
 
 
-@pytest.mark.parametrize('field_class', [ftr_fields.Body, ftr_fields.Path])
-def test_bad_multi_field(field_class):
-    with pytest.raises(ftr_errors.InvalidParameterTypeError):
-        field_class(multi=True)
+@pytest.mark.parametrize('annotation, expected', [
+    (str, False),
+    (list, True),
+    (list[str], True),
+    (typing.List[str], True), # noqa UP006
+    (set, True),
+    (set[str], True),
+    (typing.Set[str], True), # noqa UP006
+    (frozenset, True),
+    (frozenset[str], True),
+    (typing.FrozenSet[str], True), # noqa UP006
+    (tuple, True),
+    (tuple[str], True),
+    (typing.Tuple[str], True), # noqa UP006
+    (typing.Annotated[str, pydantic.Field(alias='field')], False),
+    (typing.Annotated[list[str], pydantic.Field(alias='list-field')], True),
+    (typing.Annotated[typing.List[str], pydantic.Field(alias='list-field')], True), # noqa UP006
+])
+def test_field_is_multi_field(annotation, expected):
+    assert ftr_fields.Field.is_multi_field(annotation) == expected